Change constructors
The default constructors previously allocated 0 on the lists. That was a
mistake because if a graph was instantiated using the default
constructor the user most likely wanted to create a graph with no
specific expected number of vertices. The default constructor will now
instantiate the lists with the default capacity which seems more
reasonable.